requirements.txt
requests==2.7.0
robotframework==3.2.1
docutils==0.16